renames interface.py to routes.py and refactors code
